CarDashboard/app/requirements.txt
pillow
opencv-python
numpy
scipy
matplotlib
scikit-image
scikit-learn
ipython
Face-recognition
firebase_admin
ytmusicapi
youtube_dl
pafy
haversine